Adjusting graphql schema totalDistance field to be Float type

+++ analytics/core/gql_schema.py
@@ -17,7 +17,7 @@ type_defs = gql("""
     type Exercise {
         exerciseType: ExerciseType!
         totalDuration: Int!
-        totalDistance: Int
+        totalDistance: Float
         averagePace: Float
         averageSpeed: Float
         topSpeed: Float
